fix: split the cli into modules
